fix: Apply code review feedback
